Clean up
Remove obsolete code, optimize imports etc.
