+++ modules/builders/qmstr-gradle-plugin/src/main/java/org/qmstr/gradle/QmstrPlugin.java
@@ -2,17 +2,7 @@ package org.qmstr.gradle;
 
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
-import org.gradle.api.distribution.plugins.DistributionPlugin;
-import org.gradle.api.plugins.JavaPlugin;
-import org.gradle.api.plugins.JavaPluginConvention;
-import org.gradle.api.tasks.SourceSet;
-
 import org.qmstr.gradle.android.QmstrAndroidAction;
-
-import java.util.Collections;
-import java.util.Set;
-
 import org.qmstr.gradle.java.QmstrJavaAction;
 
 public class QmstrPlugin implements Plugin<Project> {
@@ -21,15 +11,14 @@ public class QmstrPlugin implements Plugin<Project> {
         project.getExtensions().create("qmstr", QmstrPluginExtension.class);
         project.getLogger().lifecycle("Applying QMSTR plugin to {}!", project.getName());
 
-
         // Apply plugin for all java subprojects
         project.getAllprojects().stream()
             .forEach(pro -> {
                 pro.getPluginManager().apply(QmstrPlugin.class);
             });
 
-            project.getPluginManager().withPlugin("java", new QmstrJavaAction(project));
-            project.getPluginManager().withPlugin("com.android.library", new QmstrAndroidAction(project));
-            project.getPluginManager().withPlugin("com.android.application", new QmstrAndroidAction(project));
+        project.getPluginManager().withPlugin("java", new QmstrJavaAction(project));
+        project.getPluginManager().withPlugin("com.android.library", new QmstrAndroidAction(project));
+        project.getPluginManager().withPlugin("com.android.application", new QmstrAndroidAction(project));
     }
 }
